update begin and finish migration to only work on the primary range of the node nodetool was called on
